pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pybeckerplus"
version = "0.1.0"
description = "A Python library for the Becker Centronic Plus USB protocol"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "honsma235",email = "114189750+honsma235@users.noreply.github.com"}
]
dependencies = [
    "pyserial-asyncio",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Home Automation",
    "Intended Audience :: Developers",
]